refactor(database): split tariff DAO functions into tariffDao.go (#25)
Pure file-organization refactor. Moves all six Tariff functions plus
ErrTariffUtilized from database.go into a new tariffDao.go. Matches
the structure that vfeeg-backend v0.3.05 (production branch) uses —
see ADR-0004 §"Tariff-DAO-Separation".

Functions moved (unchanged):
- GetTariff(dbOpen, tenant)
- GetTariffHistory(dbOpen, tenant, id)
- GetTariffNameMap(dbOpen, tenant)
- ArchiveTariff(dbConn, tenant, id)
- AddTariff(dbConn, tenant, tariff)
- UpdateTariff(tenant, tariff)
- ErrTariffUtilized (sentinel)

database.go retains only the connection helpers (GetDBConnection,
GetDBXConnection) and the package-level pgDialect var — net -175 LoC
moved, +183 LoC added (the difference is the new file's package +
import block, copyright-equivalent).

No behavior change. go test ./... 8/8 packages green.
